Created the TaskAssigner script
Created the TaskAssigner script to assign tasks to the player and allow their "completion" by clicking the mouse. To view the tasks and their completion please view the console. This was all done on the TestShip scene, where the script is applied to the player, so this SHOULD NOT work in any other scene. Currently also working on adding UI text displaying the current tasks.