Added foundation for body reporting and voting
Added a PlayerUI script to hold UI code. Added a Body script to hold data for a dead body. Added a sample dead body on the TestShip scene with functional UI telling the player to report when they walk over it. Added a placeholder screen for future voting screen functionality. Currently when the voting screen is opened there is no way to close it except by stopping the game and relaunching in the editor.