Camera/Map/Sprite/Art Changes
Extended Test Ship and added thicker boundaries.
Added Player Sprite for Testing (His name is Bobby) The sprite is attached to the player. The Physical pirate body is still there just toggle the renderer back on to see it.
Adjusted Camera angle a little.
Added a new folder under Assets called Art and moved materials inside this folder. Now, textures can be added directly into this folder and any new materials can be added to the materials folder inside Art folder.

Note: The sprite still rotates with WASD. This will be altered in the future (the sprite disappears when parallel the camera).