Added Helm
We have a helm now!
+Helm Model (Colored and Blank)
+Helm UV Map
+Helm Prefab

The wheel is free and separate from the stand. This should allow for whoever is working on the helm task to animate the rotation. I've provided two separate versions (one colored from a template and a blank version) The UV mesh is also in the folder.

Feel free to delete/change/add anything to the model