Added a Capstan
Added a capstan model with simple texture.

Capstan Definition: A large (cylindrical) mechanical device with lever arms. On ships, crews would work together to push/turn the capstan in order to lift heavy loads/ and or anchors.

For the purpose of this game, it could lift/lower an anchor or two... this would be a good visual task to help prove that you're not sus.