Added basic voting functionality.
Should now be able to initiate a vote over the network, allow every player to vote on it. The vote only ends when every player has voted. Once this happens the player who has been voted out will display in the console. Still need to add a way for the server to automatically instantiate an appropriate amount of voting nodes for each player. There are currently some errors that show up in the console due to issues with some of the RPCs calling on photonViews that they shouldn't. This is expected and, as of my minor testing, does not affect gameplay. Additionally, multiple back-to-back votes should not cause any problems. Something to look into: Might be issues with task UI messing up after a vote, need to check further.