Added a really basic main menu UI.
Shouldn't have any issues far as I know. 2 things to note tho: 1. If you hit the join or create match buttons too early (before PUN connects to our server) it'll cause an error. Nothing to worry about tho, as if you just wait a sec and press the button again it'll work normally. 2. The quit button will not function if you run the program in Unity Engine, it will only work on built versions of the game, which is by Unity's design, so dont worry if that button does not function.