Tried to create some beginner lobby functions
So I tried to make beginner lobby functionality. Far as I know everything works with the exception of the ready up feature. To ready up all players must press E on their keyboards, which will activate all of the scripts (task assigner and UI) that are needed for base gameplay. The only problem is when I tested this with two versions of Unity open, every time one pressed E, both players were ready, instead of the expected one. Unless I'm missing something about Unity input this should not be the case, as I thought input from a machine would only apply to one open instance of the game at a time. If I can get home to my PC I'll be able to do more extensive testing, but for now this is what we got.