Ready up kinda works now
Ready up mechanic moved from being a keyboard press to being a clickable button on the screen. With this change, the previous issues with ready up are fixed. Next, need to work on moving the ready up button, making the text on it change to say unready when readied, having a number of readied people text, and make it go away at appropriate times.