More ready button tweaks
Now the ready button is accompanied by a text object above it that says how many players are in the lobby and how many are ready. With this change, the ready button generally seems to be functional. Now planning to move the ready button to somewhere more reasonable and making a little lobby area to make the beginning of the game actually functional with an opening sequence when the game starts and everything. Should be hella cool :)